Create pluto.c
pluto implementation of a 3d wave equation. 
This code has been taken from PLUTO's repository at http://repo.or.cz/w/pluto.git
It has been slightly modified in order to accept input parameters from a file instead of command line.
In order to let it work with PLUTO is needed to apply a post-compilation script, available on their website and slightly modified in order to let it work on our machine. It's name is replace.sh, also available in this repository
Please for any issue refer to them