Create replace.sh
Post-compilation script to be executed in order for the pluto's generated code to work.
The original is available at http://repo.or.cz/w/pluto.git. For any issue refer to them.
We slightly modified it in order to let it work with any(?) version of the GCC we tried (and not be bound to a specific version and machine).